ICE-6253 Functional_TwowayCallbackDouble is invalid

+++ java/src/Ice/src/main/java/IceInternal/Functional_TwowayCallbackDouble.java
@@ -30,7 +30,7 @@ public abstract class Functional_TwowayCallbackDouble
         __responseCb = responseCb;
     }
 
-    public void response(byte arg)
+    public void response(double arg)
     {
         if(__responseCb != null)
         {
